Move predefined key loader to common4j from common
